requirements.txt
fastapi==0.95.1
uvicorn==0.22.0

#new
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6

python-dotenv==1.0.0
python-multipart==0.0.9
alembic==1.13.1
boto3
